Fixes & Enhancements: RaidStart Evidence Trigger, Ping Hotkey Fix, Playercard State Sync, Prayer Synchronization
• Raid Start Evidence Fix (Walking Around Issue): Replaced fragile lobby Region ID checks with precise instanced raid detection (`isPlayerInRaid()`). Auto-ready resets, raid rules printing, and auto-screenshots now ONLY trigger when actually entering a raid instance, stopping unwanted triggers when players simply walk around.

• Instant State Sync: Eliminated the delay on party panel updates. Local health, prayer, stats, and equipment changes are now tracked tick-by-tick and broadcast immediately across the party WebSocket, enabling real-time prayer flicking responsiveness.

• Chatbox Ping Hotkey Hijacking Fix: Resolved an issue where pressing ping hotkeys (e.g., 'R') while typing messages in the chatbox would trigger pings and swallow keystrokes when typing words containing that hotkey. **You don't have to use "Press Enter to Chat..." - You just cannot begin typing with your hotkey**

• Royal Titans Prayer Scroll Separation: Added standalone party card icons for Eagle Eye vs. Deadeye and Mystic Might vs. Mystic Vigour.

• Scroll State Synchronization: Synchronized internal Varbit and widget tree inspection logic so accounts with Royal Titans scrolls properly broadcast upgraded prayers (Might/Eagle Eye) while base accounts broadcast base prayers (Vigour/Deadeye).

• Minimap Visual Overlay: Added `RaidPartyMinimapOverlay` to render party members onto the minimap.

• Ghost Member Disconnect Cleanup: Added automated heartbeat monitoring to purge lingering "ghost" player cards when party members close their clients or disconnect unexpectedly.

**Check out the settings panel on the new update!**

+++ src/main/java/com/github/kalsupes/raidparty/RaidPartyPlayerSync.java
@@ -37,7 +37,7 @@ public class RaidPartyPlayerSync extends PartyMemberMessage {
     private int readyState;
 
     // Active prayer IDs (ordinals packed into bits)
-    private int activePrayers;
+    private long activePrayers;
 
     // Skill levels: [boosted, real] for each Skill ordinal
     // Index = Skill.ordinal() * 2 (boosted), Skill.ordinal() * 2 + 1 (real)
@@ -67,8 +67,8 @@ public class RaidPartyPlayerSync extends PartyMemberMessage {
     private int quiverAmmoQty; // VarPlayer.DIZANAS_QUIVER_ITEM_COUNT
 
     // Prayer states (ordinals packed into bits)
-    private int availablePrayers; // Prayers player has the level for
-    private int unlockedPrayers; // Prayers unlocked via quests/rewards
+    private long availablePrayers; // Prayers player has the level for
+    private long unlockedPrayers; // Prayers unlocked via quests/rewards
 
     // Spellbook: 0=Standard, 1=Ancient, 2=Lunar, 3=Arceuus
     private int spellbook;
